Implemented operator support
Implemented support for all of the basic operators this calculator
will support. These are  '+', '-', '*', '/', 'sq' (square),
'sqrt' (square root), 'dup' (duplicate), 'swap', 'ps' (print stack),
and 'quit'.

Note that this is an UNSTABLE commit; the project builds without
error and runs without crashing, but there has been NO bug hunting
or testing. This is most likely WRONG, but it's better than it was.
Maybe.

* main.cpp - Implemented all operators.

* deliverables/log.md - Documented changes.