Fixed assorted bugs
Fixed assorted bugs with the calculator programs.

The first main issue was that a 0 appeard at the top of the stack
whenever the program was launched. I traced this to a bug in the
Stack.peek() code, where a 0 was generated and pushed if we tried
to peek at an empty stack. This was a pretty easy fix, a few checks
against Stack.getSize() and everything was golden.

The second major issue was the inability to use the '-' operator,
as anything starting with a '-' was interpreted as a negative
operand. By checking the strlen() of the input string, I was able
to differentiate between "-" (length of 1), and "-n" (length of
something other than 1).

The final issue was various formatting tweaks necessary to match
the required format for the assignment.

* main.cpp - Fixed '-' bug. Fixed formatting inconsistencies.

* stack.cpp - Fixed Stack.pop() bug.